cosmetics
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@7238 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/sphinx4/edu/cmu/sphinx/util/MatrixUtils.java
@@ -98,7 +98,7 @@ public class MatrixUtils {
 
 
     /** Converts a vector from linear domain to logdomain using a given <code>LogMath</code>-instance for conversion. */
-    public static float[] linearToLog(double[] vector, LogMath logMath) {
+    public static float[] linearToLog(float[] vector, LogMath logMath) {
         float[] logMixtureWeights = new float[vector.length];
         int nbGaussians = vector.length;
         for (int i = 0; i < nbGaussians; i++) {
